Hayyi

+++ database/migrations/2024_02_02_035138_create_sale_documents_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('code_document_sale')->unique();
             $table->string('buyer_name_document_sale');
             $table->bigInteger('total_product_document_sale');
-            $table->bigInteger('total_price_document_sale');
+            $table->decimal('total_price_document_sale', 15, 2);
             $table->enum('status_document_sale', ['proses', 'selesai']);
             $table->timestamps();
         });
